separate res file generation in another class
